+++ heft/Shared/Frame.h
@@ -33,7 +33,7 @@ public:
 	static const int ciMinSize = 6;
 
 	Frame(const UINT8* pData, int len, bool partial);
-	int GetLength(){return write_data.size();}
+	int GetLength(){return (int)write_data.size();}
 	UINT8* GetData(){return &write_data[0];}
     static int GetMetaDataSize(){return ciMinSize;}
